[Tests] Re-enable \Bolt\Tests\Asset\SnippetsTest
